requirements.txt
Flask==2.3.2
pydantic_core==2.3.0
